support ssh-agent on windows
support ssh-agent on windows named pipe \\.\pipe\open-ssh-agent
