+++ tssh/login.go
@@ -45,7 +45,6 @@ import (
 	"github.com/skeema/knownhosts"
 	"github.com/trzsz/trzsz-go/trzsz"
 	"golang.org/x/crypto/ssh"
-	"golang.org/x/crypto/ssh/agent"
 	"golang.org/x/term"
 )
 
@@ -535,35 +534,6 @@ var getDefaultSigners = func() func() []*sshSigner {
 		return signers
 	}
 }()
-
-func getAgentSigners() []*sshSigner {
-	socket := os.Getenv("SSH_AUTH_SOCK")
-	if socket == "" {
-		return nil
-	}
-
-	conn, err := net.Dial("unix", socket)
-	if err != nil {
-		debug("open ssh agent [%s] failed: %v", socket, err)
-		return nil
-	}
-	cleanupAfterLogined = append(cleanupAfterLogined, func() {
-		conn.Close()
-	})
-
-	client := agent.NewClient(conn)
-	signers, err := client.Signers()
-	if err != nil {
-		debug("get ssh agent signers failed: %v", err)
-		return nil
-	}
-
-	wrappers := make([]*sshSigner, 0, len(signers))
-	for _, signer := range signers {
-		wrappers = append(wrappers, &sshSigner{path: "ssh-agent", pubKey: signer.PublicKey(), signer: signer})
-	}
-	return wrappers
-}
 
 func getPublicKeysAuthMethod(args *sshArgs) ssh.AuthMethod {
 	if strings.ToLower(getOptionConfig(args, "PubkeyAuthentication")) == "no" {
